Moved old types to v2
